Claims downloading has been implemented and features drawing improvements made

+++ DMI/src/main/java/com/rmsi/mast/studio/dao/RelationshipTypeDao.java
@@ -2,6 +2,6 @@ package com.rmsi.mast.studio.dao;
 
 import com.rmsi.mast.studio.domain.RelationshipType;
 
-public interface RelationshipTypeDao extends GenericDAO<RelationshipType, Integer> {
+public interface RelationshipTypeDao extends GenericDAO<RelationshipType, Long> {
     RelationshipType getTypeByAttributeOptionId(int optId);
 }
